Mutualize kara folder creation for integration tests
